move sessions list to separate file
